[FE][Bom-239] fix: 하이라이트 생성 수정 삭제할 때 하이라이트가 깜빡이는 이슈 (#191)
* refactor: highlight api query factory 적용

* fix: type 오류 해결

* refactor: ArticleContent 컴포넌트 분리

* fix: window selection 수동 제거

* fix: 안쓰는 로직 제거

* chore: console 삭제

* refactor: 타입 수정
