+++ frontend/src/pages/detail/components/FloatingToolbar/FloatingToolbar.tsx
@@ -55,6 +55,7 @@ export default function FloatingToolbar({
       selection: selectionRef.current,
       highlightId: selectedHighlightId,
     });
+    window.getSelection()?.removeAllRanges();
   };
 
   const handleMemoClick = () => {
@@ -63,6 +64,7 @@ export default function FloatingToolbar({
       mode: currentMode,
       selection: selectionRef.current,
     });
+    window.getSelection()?.removeAllRanges();
   };
 
   useEffect(() => {
@@ -107,8 +109,8 @@ export default function FloatingToolbar({
       hideToolbar();
     };
 
-    document.addEventListener('click', handleDocumentClick);
-    return () => document.removeEventListener('click', handleDocumentClick);
+    document.addEventListener('mouseup', handleDocumentClick);
+    return () => document.removeEventListener('mouseup', handleDocumentClick);
   }, [selectionTargetRef]);
 
   return (
