+++ frontend/src/routes/_bombom/articles.$articleId.tsx
@@ -9,7 +9,7 @@ import Chip from '@/components/Chip/Chip';
 import Spacing from '@/components/Spacing/Spacing';
 import useScrollRestoration from '@/hooks/useScrollRestoration';
 import { useScrollThreshold } from '@/hooks/useScrollThreshold';
-import ArticleContent from '@/pages/detail/components/ArticleContent/ArticleContent';
+import ArticleBody from '@/pages/detail/components/ArticleBody/ArticleBody';
 import EmptyUnreadCard from '@/pages/detail/components/EmptyUnreadCard/EmptyUnreadCard';
 import FloatingActionButtons from '@/pages/detail/components/FloatingActionButtons/FloatingActionButtons';
 import NewsletterItemCard from '@/pages/detail/components/NewsletterItemCard/NewsletterItemCard';
@@ -86,7 +86,7 @@ function ArticleDetailPage() {
         </MetaInfoRow>
       </HeaderWrapper>
       <Divider />
-      <ArticleContent
+      <ArticleBody
         articleId={articleIdNumber}
         articleContent={currentArticle.contents}
       />
